project2/face-recognition-lambda/requirements.txt
boto3
torch
torchvision
facenet-pytorch
opencv-python-headless
pillow
imutils
numpy